+++ dorm_annotations/example/lib/drawing.dorm.dart
@@ -240,3 +240,14 @@ class Dorm<Q extends BaseQuery<Q>, P extends PageRequest> {
   >
   get drawings => DatabaseEntity(const DrawingEntity(), engine: _engine);
 }
+
+class TransactionalDorm<Q extends BaseQuery<Q>, P extends PageRequest>
+    extends Dorm<Q, P> {
+  const TransactionalDorm(this._transactionalEngine)
+    : super(_transactionalEngine);
+
+  final TransactionalEngine<Q, P> _transactionalEngine;
+
+  Future<T> transaction<T>(Future<T> Function(Dorm<Q, P>) action) =>
+      _transactionalEngine.transaction((engine) => action(Dorm<Q, P>(engine)));
+}