+++ dorm_bloc_database/test/compliance_test.dart
@@ -2,7 +2,7 @@ import 'package:dorm_bloc_database/dorm_bloc_database.dart';
 import 'package:dorm_framework/dorm_framework.dart';
 import 'package:dorm_test/dorm_test.dart';
 
-class _BlocSession implements EngineTestSession<Query> {
+class _BlocSession implements TransactionalEngineTestSession<Query> {
   _BlocSession() : _engine = Engine();
 
   final Engine _engine;
@@ -11,9 +11,14 @@ class _BlocSession implements EngineTestSession<Query> {
   BaseEngine<Query, OffsetPageRequest> get engine => _engine;
 
   @override
+  TransactionalEngine<Query, OffsetPageRequest> get transactionalEngine =>
+      _engine;
+
+  @override
   final EngineCapabilities capabilities = const EngineCapabilities(
     compositeIdentities: true,
     reactiveStreams: true,
+    transactions: true,
   );
 
   @override
@@ -23,14 +28,15 @@ class _BlocSession implements EngineTestSession<Query> {
   Future<void> close() async {}
 }
 
-class _BlocAdapter implements EngineTestAdapter<Query> {
+class _BlocAdapter implements TransactionalEngineTestAdapter<Query> {
   @override
   String get name => 'BLoC';
 
   @override
-  Future<EngineTestSession<Query>> open() async => _BlocSession();
+  Future<TransactionalEngineTestSession<Query>> open() async => _BlocSession();
 }
 
 void main() {
   defineEngineComplianceTests(_BlocAdapter());
+  defineEngineTransactionComplianceTests(_BlocAdapter());
 }