+++ dorm_memory_database/test/compliance_test.dart
@@ -2,7 +2,7 @@ import 'package:dorm_framework/dorm_framework.dart';
 import 'package:dorm_memory_database/dorm_memory_database.dart';
 import 'package:dorm_test/dorm_test.dart';
 
-class _MemorySession implements EngineTestSession<Query> {
+class _MemorySession implements TransactionalEngineTestSession<Query> {
   _MemorySession() : _engine = Engine();
 
   final Engine _engine;
@@ -11,9 +11,14 @@ class _MemorySession implements EngineTestSession<Query> {
   BaseEngine<Query, OffsetPageRequest> get engine => _engine;
 
   @override
+  TransactionalEngine<Query, OffsetPageRequest> get transactionalEngine =>
+      _engine;
+
+  @override
   final EngineCapabilities capabilities = const EngineCapabilities(
     compositeIdentities: true,
     reactiveStreams: true,
+    transactions: true,
   );
 
   @override
@@ -23,14 +28,16 @@ class _MemorySession implements EngineTestSession<Query> {
   Future<void> close() async {}
 }
 
-class _MemoryAdapter implements EngineTestAdapter<Query> {
+class _MemoryAdapter implements TransactionalEngineTestAdapter<Query> {
   @override
   String get name => 'Memory';
 
   @override
-  Future<EngineTestSession<Query>> open() async => _MemorySession();
+  Future<TransactionalEngineTestSession<Query>> open() async =>
+      _MemorySession();
 }
 
 void main() {
   defineEngineComplianceTests(_MemoryAdapter());
+  defineEngineTransactionComplianceTests(_MemoryAdapter());
 }